TH-17 Rename 'DisplayImage' component as 'DisplayPersonImage
