+++ client/assets/emptyPersonImage.png
