+++ client/components/Molecules/DisplayPersonImage/styledComponents.ts
