+++ client/components/Molecules/index.ts
@@ -1,6 +1,6 @@
 export { default as CardDescription } from "./CardDescription";
 export { default as CardTitle } from "./CardTitle";
-export { default as DisplayImage } from "./DisplayImage";
+export { default as DisplayPersonImage } from "./DisplayPersonImage";
 export { default as FloatingOpenMenu } from "./FloatingOpenMenu";
 export { default as FormField } from "./FormField";
 export { default as IndexPageTitle } from "./IndexPageTitle";
